Change Fragment x and y type

+++ src/Fragment.cpp
@@ -1,6 +1,6 @@
 #include "Fragment.h"
 
-Fragment::Fragment(int x, int y, double z) {
+Fragment::Fragment(size_t x, size_t y, double z) {
 	this->x = x;
 	this->y = y;
 	this->z = z;
